requirements.txt
streamlit
pypdf2
python-dotenv
faiss-cpu
openai
huggingface_hub

# LangChain ecosystem
langchain
langchain-community
langchain-openai
langchain-huggingface

# For HuggingFace embeddings/models
sentence-transformers
